fix: Split ImageBubble into platform-specific files (iOS/Android)
