+++ package.json
@@ -39,6 +39,7 @@
     "@chatwoot/utils": "^0.0.33",
     "@gorhom/bottom-sheet": "^5.1.2",
     "@kesha-antonov/react-native-action-cable": "^1.1.5",
+    "@nandorojo/galeria": "^1.2.0",
     "@notifee/react-native": "^9.1.1",
     "@react-native-async-storage/async-storage": "^1.23.1",
     "@react-native-clipboard/clipboard": "^1.6.1",
