+++ src/hooks/useImageDimensions.ts
@@ -1,27 +1,22 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { Image as RNImage } from 'react-native';
-import ImageModal from 'react-native-image-modal';
-import { tailwind } from '@/theme';
 
-type ImageCellProps = {
+export type ImageCellProps = {
   imageSrc: string;
 };
 
-type ImageContainerProps = Pick<ImageCellProps, 'imageSrc'> & {
+export type ImageContainerProps = Pick<ImageCellProps, 'imageSrc'> & {
   maxWidth?: number;
   maxHeight?: number;
 };
 
 const imageDimensionsCache = new Map<string, { width: number; height: number }>();
 
-/**
- * Renders a chat image preview and fullscreen modal.
- *
- * The preview size is derived from the original image resolution and
- * constrained by maxWidth/maxHeight while preserving aspect ratio.
- */
-export const ImageBubbleContainer = (props: ImageContainerProps) => {
-  const { imageSrc, maxWidth = 300, maxHeight = 360 } = props;
+export const useImageDimensions = (
+  imageSrc: string,
+  maxWidth = 300,
+  maxHeight = 360,
+) => {
   const [imageDimensions, setImageDimensions] = useState<{ width: number; height: number } | null>(
     null,
   );
@@ -78,25 +73,5 @@ export const ImageBubbleContainer = (props: ImageContainerProps) => {
     return { width, height };
   }, [imageDimensions, maxHeight, maxWidth]);
 
-  return (
-    <ImageModal
-      source={{ uri: imageSrc }}
-      resizeMode="contain"
-      modalImageResizeMode="contain"
-      overlayBackgroundColor="#000000"
-      imageBackgroundColor="#F3F4F6"
-      isTranslucent
-      style={[tailwind.style('bg-gray-100 overflow-hidden'), imageStyle]}
-    />
-  );
-};
-
-export const ImageBubble = (props: ImageCellProps) => {
-  const { imageSrc } = props;
-
-  return (
-    <React.Fragment>
-      <ImageBubbleContainer {...{ imageSrc }} />
-    </React.Fragment>
-  );
+  return imageStyle;
 };
